First commit of project 13

+++ controllers/clothingItems.js
@@ -1,8 +1,9 @@
-const ClothingItem = require('../models/clothingItem');
+const ClothingItem = require("../models/clothingItem");
 const {
   BAD_REQUEST_STATUS_CODE,
   NOT_FOUND_STATUS_CODE,
   SERVER_ERROR_STATUS_CODE,
+  FORBIDDEN_STATUS_CODE,
 } = require("../utils/errors");
 
 module.exports.getItems = (req, res) => {
@@ -14,9 +15,9 @@ module.exports.getItems = (req, res) => {
         .status(SERVER_ERROR_STATUS_CODE)
         .send({ message: "An error has occurred on the server" });
     });
-}
+};
 
- module.exports.createItem = (req, res) => {
+module.exports.createItem = (req, res) => {
   const { name, weather, imageUrl } = req.body;
   const owner = req.user._id;
   ClothingItem.create({ name, weather, imageUrl, owner })
@@ -72,10 +73,23 @@ const updateLike = (req, res, method) => {
 };
 
 module.exports.deleteItem = (req, res) => {
+  const userId = req.user._id;
   const { itemId } = req.params;
-  ClothingItem.findByIdAndDelete(itemId)
-    .orFail()
-    .then((item) => res.status(200).send(item))
+  ClothingItem.findById(itemId)
+    .orFail(() => {
+      const error = new Error("Item not found");
+      error.statusCode = NOT_FOUND_STATUS_CODE;
+      throw error;
+    })
+    .then((item) => {
+      if (item.owner.toString() !== userId.toString()) {
+        const error = new Error("You are not authorized to delete this item");
+        error.statusCode = FORBIDDEN_STATUS_CODE;
+        throw error;
+      }
+      return ClothingItem.findByIdAndDelete(itemId)
+        .then(() => res.json({ message: "Item successfully deleted" }));
+    })
     .catch((err) => {
       console.error(err);
       if (err.name === "ValidationError") {
